refactor: move startXSnap to new file
